updated time types

+++ src/main.cpp
@@ -49,14 +49,14 @@ void setup() {
     delay(3000);
 
     // Loop runs for calculated time required to complete three full revolutions, plus a n additional amount
-    float t_init {millis()};
-    float t_out {millis()};
+    long unsigned int t_init {millis()};
+    long unsigned int t_out {millis()};
     while ((t_out - t_init) < t_req * 1000 * 3 * 0.5 + 2000)
     {
         // Inner loop runs on x fraction of t_interval, moving forward for x*t_interval and backward for (1-x)*t_interval
         float t_interval {250};
 
-        float t_start {millis()};
+        long unsigned int t_start {millis()};
         float t = t_start;
         while ((t - t_start) < t_interval * x)
         {
@@ -67,7 +67,7 @@ void setup() {
             t = millis();
         }
 
-        float t_start2 {millis()};
+        long unsigned int t_start2 {millis()};
         t = t_start2;
         while ((t - t_start2) < t_interval * (1 - x))
         {
